msnew(memstack*,int) => msnew(memstack*,size_t) to me more like stdlib

+++ include/memstack.h
@@ -56,7 +56,7 @@ typedef struct {
 
 extern void msinit(); // Initializes a global memstack, that can be accessed by passing in "NULL" as storage in the other funcs
 extern memstack* msnew();  // Creates a new memstack
-extern void* msalloc(memstack* storage, int size);  // Allocate onto memstack and return pointer
+extern void* msalloc(memstack* storage, size_t size);  // Allocate onto memstack and return pointer
 extern void msfree(memstack* storage);  // Free all of memstack
 extern void msclear(memstack* storage);  // Frees allocations but doesn't delete the memstack
 extern void mspush(memstack* storage, void* ptr);  // Adds new memory into the memstack even if it hasn't been allocated with msalloc()
